+++ app/src/main/java/com/growit/app/retrospect/controller/goalretrospect/dto/GoalRetrospectResponse.java
@@ -1,4 +1,4 @@
-package com.growit.app.goal.controller.goalretrospect.dto;
+package com.growit.app.retrospect.controller.goalretrospect.dto;
 
 import com.growit.app.retrospect.domain.goalretrospect.vo.Analysis;
 import lombok.AllArgsConstructor;
