+++ app/src/main/java/com/growit/app/retrospect/controller/retrospect/dto/request/UpdateRetrospectRequest.java
@@ -1,4 +1,4 @@
-package com.growit.app.retrospect.controller.dto.request;
+package com.growit.app.retrospect.controller.retrospect.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -7,13 +7,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CreateRetrospectRequest {
-  @NotBlank(message = "{validation.retrospect.goal-id.required}")
-  private String goalId;
-
-  @NotBlank(message = "{validation.retrospect.plan-id.required}")
-  private String planId;
-
+public class UpdateRetrospectRequest {
   @NotBlank(message = "{validation.retrospect.content.required}")
   @Size(min = 10, max = 200, message = "{validation.retrospect.content.size}")
   private String content;
