+++ app/src/main/java/com/growit/app/retrospect/controller/retrospect/dto/response/RetrospectResponse.java
@@ -1,4 +1,4 @@
-package com.growit.app.retrospect.controller.dto.response;
+package com.growit.app.retrospect.controller.retrospect.dto.response;
 
 import com.growit.app.goal.domain.goal.plan.Plan;
 import lombok.AllArgsConstructor;
