+++ app/src/main/java/com/growit/app/retrospect/controller/retrospect/mapper/RetrospectResponseMapper.java
@@ -1,8 +1,8 @@
-package com.growit.app.retrospect.controller.mapper;
+package com.growit.app.retrospect.controller.retrospect.mapper;
 
-import com.growit.app.retrospect.controller.dto.response.RetrospectExistResponse;
-import com.growit.app.retrospect.controller.dto.response.RetrospectResponse;
-import com.growit.app.retrospect.usecase.dto.RetrospectWithPlan;
+import com.growit.app.retrospect.controller.retrospect.dto.response.RetrospectExistResponse;
+import com.growit.app.retrospect.controller.retrospect.dto.response.RetrospectResponse;
+import com.growit.app.retrospect.usecase.retrospect.dto.RetrospectWithPlan;
 import org.springframework.stereotype.Component;
 
 @Component
