+++ app/src/main/java/com/growit/app/retrospect/usecase/retrospect/GetRetrospectByFilterUseCase.java
@@ -1,4 +1,4 @@
-package com.growit.app.retrospect.usecase;
+package com.growit.app.retrospect.usecase.retrospect;
 
 import com.growit.app.goal.domain.goal.Goal;
 import com.growit.app.goal.domain.goal.plan.Plan;
@@ -6,7 +6,7 @@ import com.growit.app.goal.domain.goal.service.GoalQuery;
 import com.growit.app.retrospect.domain.retrospect.Retrospect;
 import com.growit.app.retrospect.domain.retrospect.dto.RetrospectQueryFilter;
 import com.growit.app.retrospect.domain.retrospect.service.RetrospectQuery;
-import com.growit.app.retrospect.usecase.dto.RetrospectWithPlan;
+import com.growit.app.retrospect.usecase.retrospect.dto.RetrospectWithPlan;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
