+++ app/src/main/java/com/growit/app/retrospect/usecase/retrospect/UpdateRetrospectUseCase.java
@@ -1,4 +1,4 @@
-package com.growit.app.retrospect.usecase;
+package com.growit.app.retrospect.usecase.retrospect;
 
 import com.growit.app.retrospect.domain.retrospect.Retrospect;
 import com.growit.app.retrospect.domain.retrospect.RetrospectRepository;
