+++ app/src/test/java/com/growit/app/goal/controller/goalretrospect/GoalRetrospectControllerTest.java
@@ -15,11 +15,11 @@ import com.epages.restdocs.apispec.ResourceSnippetParametersBuilder;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.growit.app.common.TestSecurityUtil;
 import com.growit.app.fake.goalretrospect.GoalRetrospectFixture;
-import com.growit.app.goal.controller.goalretrospect.dto.UpdateGoalRetrospectRequest;
+import com.growit.app.retrospect.controller.goalretrospect.dto.UpdateGoalRetrospectRequest;
 import com.growit.app.retrospect.domain.goalretrospect.GoalRetrospect;
 import com.growit.app.retrospect.domain.goalretrospect.vo.Analysis;
-import com.growit.app.retrospect.usecase.GetGoalRetrospectUseCase;
-import com.growit.app.retrospect.usecase.UpdateGoalRetrospectUseCase;
+import com.growit.app.retrospect.usecase.goalretrospect.GetGoalRetrospectUseCase;
+import com.growit.app.retrospect.usecase.goalretrospect.UpdateGoalRetrospectUseCase;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
