+++ app/src/test/java/com/growit/app/retrospect/usecase/CheckRetrospectExistsByPlanIdUseCaseTest.java
@@ -8,6 +8,7 @@ import com.growit.app.fake.retrospect.RetrospectFixture;
 import com.growit.app.goal.domain.goal.service.GoalValidator;
 import com.growit.app.retrospect.domain.retrospect.dto.RetrospectQueryFilter;
 import com.growit.app.retrospect.domain.retrospect.service.RetrospectQuery;
+import com.growit.app.retrospect.usecase.retrospect.CheckRetrospectExistsByPlanIdUseCase;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
