+++ app/src/test/java/com/growit/app/retrospect/usecase/GetRetrospectUseCaseTest.java
@@ -10,7 +10,8 @@ import com.growit.app.goal.domain.goal.service.GoalQuery;
 import com.growit.app.retrospect.domain.retrospect.Retrospect;
 import com.growit.app.retrospect.domain.retrospect.dto.GetRetrospectQueryFilter;
 import com.growit.app.retrospect.domain.retrospect.service.RetrospectQuery;
-import com.growit.app.retrospect.usecase.dto.RetrospectWithPlan;
+import com.growit.app.retrospect.usecase.retrospect.GetRetrospectUseCase;
+import com.growit.app.retrospect.usecase.retrospect.dto.RetrospectWithPlan;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
