+++ app/src/test/java/com/growit/app/retrospect/usecase/GetRetrospectsByGoalIdUseCaseTest.java
@@ -11,7 +11,8 @@ import com.growit.app.goal.domain.goal.plan.Plan;
 import com.growit.app.goal.domain.goal.service.GoalQuery;
 import com.growit.app.retrospect.domain.retrospect.Retrospect;
 import com.growit.app.retrospect.domain.retrospect.service.RetrospectQuery;
-import com.growit.app.retrospect.usecase.dto.RetrospectWithPlan;
+import com.growit.app.retrospect.usecase.retrospect.GetRetrospectsByGoalIdUseCase;
+import com.growit.app.retrospect.usecase.retrospect.dto.RetrospectWithPlan;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
