+++ app/src/test/java/com/growit/app/retrospect/usecase/UpdateRetrospectUseCaseTest.java
@@ -8,6 +8,7 @@ import com.growit.app.fake.retrospect.RetrospectFixture;
 import com.growit.app.retrospect.domain.retrospect.RetrospectRepository;
 import com.growit.app.retrospect.domain.retrospect.dto.UpdateRetrospectCommand;
 import com.growit.app.retrospect.domain.retrospect.service.RetrospectQuery;
+import com.growit.app.retrospect.usecase.retrospect.UpdateRetrospectUseCase;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
